+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+## v0.1.2
+- `affected-paths-core`, `tooling-support-*`: Add in support for composite builds being analyzed
+- `affected-paths-core`: Remove filter of root project
+- `affected-paths-core`: Fix improper project mapping for file changes in nested projects
+- `affected-paths-core`: Allow a custom Gradle installation path to be passed in to the Gradle Tooling API
+
 ## v0.1.1
 - `tooling-support`: Fix crash from `SquareProjectModelBuilder` when used on a non-Java/Android project
 - `affected-paths-core`: Adds `autoInjectPlugin` flag to `CoreOptions`, which auto-injects the "com.squareup.tooling" plugin to the build